[refactor] 로또 번호 generator 인터페이스 생성 테스트 및 구현
