+++ src/main/java/step2/LottoNumberGenerator.java
@@ -1,0 +1,5 @@
+package step2;
+
+public interface LottoNumberGenerator {
+    Lotto generate();
+}
