+++ src/test/java/step2/ProfitTest.java
@@ -1,0 +1,35 @@
+package step2;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class ProfitTest {
+
+    @Test
+    @DisplayName("수익률 구현 테스트")
+    void profit_test(){
+//        String winningNumbers = "1,2,3,4,5,6";
+//        String purchasedNumbers1 = "1,2,3,43,44,45";
+//        String purchasedNumbers2="1,2,3,4,44,45";
+//
+//        WinningLotto winningLotto = new WinningLotto(winningNumbers);
+//        List<Lotto> lottoList = new ArrayList<>();
+//        Lotto lotto1 = new Lotto(purchasedNumbers1);
+//        Lotto lotto2 = new Lotto(purchasedNumbers2);
+//        lottoList.add(lotto1);
+//        lottoList.add(lotto2);
+//
+//        Money money = new Money(5000);
+//        PurchaseManager purchaseManager = new PurchaseManager(money);
+//
+//        Lottos lottos = Lottos.of(purchaseManager);
+//
+//        Rank rank = lottos.makeStatistic(winningLotto);
+//        Profit profit = new Profit(purchaseManager,rank);
+//
+//        assertThat(profit.profit()).isEqualto(0.35);
+    }
+}
